refactor: add generic Typography component
