--- src/components/Typography/index.ts
@@ -1,3 +1,0 @@
-export * from './Header'
-
-export * from './Text'
